Refactors common into shared
